#6 구현 / FSM에 State 저장, 로드를 위해 코드 제네레이터 변경

+++ Project/Scripts/CBackgroundScript.cpp
@@ -20,7 +20,7 @@ void CBackgroundScript::LoadFromFile(ifstream& fin)
 }
 
 CBackgroundScript::CBackgroundScript()
-	: CScript(SCRIPT_TYPE::BACKGROUNDSCRIPT)
+	: CScript((UINT)SCRIPT_TYPE::BACKGROUNDSCRIPT)
 {
 }
